pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-stash"
version = "1.0.0"
description = "Version-controlled prompt stash for AI power users"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Hina", email = "maragadavekatesh@gmail.com" }
]
keywords = ["ai", "llm", "prompts", "claude", "chatgpt", "gemini", "cli", "developer-tools", "productivity", "version-control"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "gitpython>=3.1.43",
    "pyperclip>=1.9.0",
    "rapidfuzz>=3.9.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
stash = "prompt_stash.cli:main"
vault = "prompt_stash.cli:main"

[project.urls]
Homepage = "https://github.com/Venkatesh-6921/Prompt-Stash"
Repository = "https://github.com/Venkatesh-6921/Prompt-Stash"
"Bug Tracker" = "https://github.com/Venkatesh-6921/Prompt-Stash/issues"

[tool.hatch.build.targets.wheel]
packages = ["prompt_stash"]
